Split AutomationTrack into multiple files
