+++ include/AutomationTrackView.h
@@ -1,6 +1,5 @@
 /*
- * AutomationTrack.h - declaration of class AutomationTrack, which handles
- *                     automation of objects without a track
+ * AutomationTrackView.h - declaration of class AutomationTrackView
  *
  * Copyright (c) 2008-2014 Tobias Doerffel <tobydox/at/users.sourceforge.net>
  * Copyright (c) 2006-2008 Javier Serrano Polo <jasp00/at/users.sourceforge.net>
@@ -24,36 +23,21 @@
  *
  */
 
-#ifndef AUTOMATION_TRACK_H
-#define AUTOMATION_TRACK_H
+#ifndef AUTOMATION_TRACK_VIEW_H
+#define AUTOMATION_TRACK_VIEW_H
 
-#include "Track.h"
+#include "AutomationTrack.h"
+#include "TrackView.h"
 
 
-class AutomationTrack : public Track
+class AutomationTrackView : public TrackView
 {
-	Q_OBJECT
 public:
-	AutomationTrack( TrackContainer* tc, bool _hidden = false );
-	virtual ~AutomationTrack() = default;
+	AutomationTrackView( AutomationTrack* at, TrackContainerView* tcv );
+	virtual ~AutomationTrackView() = default;
 
-	virtual bool play( const TimePos & _start, const fpp_t _frames,
-						const f_cnt_t _frame_base, int _tco_num = -1 ) override;
-
-	QString nodeName() const override
-	{
-		return "automationtrack";
-	}
-
-	TrackView * createView( TrackContainerView* ) override;
-	TrackContentObject* createTCO(const TimePos & pos) override;
-
-	virtual void saveTrackSpecificSettings( QDomDocument & _doc,
-							QDomElement & _parent ) override;
-	void loadTrackSpecificSettings( const QDomElement & _this ) override;
-
-private:
-	friend class AutomationTrackView;
+	void dragEnterEvent( QDragEnterEvent * _dee ) override;
+	void dropEvent( QDropEvent * _de ) override;
 
 } ;
