+++ src/gui/widgets/TrackOperationsWidget.cpp
@@ -33,6 +33,7 @@
 
 #include "AutomationPattern.h"
 #include "AutomationTrack.h"
+#include "AutomationTrackView.h"
 #include "ColorChooser.h"
 #include "ConfigManager.h"
 #include "DataFile.h"
